+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Honeybadger 1.6.0 (unreleased) ##
 
+* Disable Rack::SSL middleware in Rails 3 test task. Rails'
+  config.force_ssl setting otherwise prevents the request from reaching
+  the controller.
+
+  *Joshua Wood*
+
 * Added a Changelog :)
 
   *Joshua Wood*